+++ Src/Common/PropertyFetcher.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
+﻿namespace Microsoft.ApplicationInsights.Common
 {
     using System;
     using System.Reflection;